+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - added gemini-2.0-flash-exp model in Agent
 - improved tool message handling within session
 - fixed overflowing of the code blocks in messages
+- pasting multiple MCP servers at once within the same JSON will add all the servers
 
 ## [0.8.0]